components/runpod_trainer/requirements.txt
runpod>=1.7.0
transformers==4.44.0
datasets>=2.16.0
accelerate>=0.30.0
boto3>=1.34.0
scikit-learn>=1.3.0
scipy>=1.11.0
safetensors>=0.4.0
PyTDC>=1.1.0
